+++ eosio_types/generate_float_type.mjs
@@ -2,10 +2,8 @@ import { GraphQLScalarType } from "graphql";
 
 /**
  * Generates a GraphQL scalar Float type of `size`.
- * @name generate_float_type
  * @param {number} size Size of the float can be 32, 64 or 128.
  * @returns {GraphQLScalarType} GraphQL scalar float type.
- * @ignore
  */
 function generate_float_type(size) {
   return new GraphQLScalarType({